pyproject.toml
[tool.poetry]
name = "secure-chat"
version = "0.1.0"
description = "A secure instant messaging application"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
PyQt6 = "^6.5.0"
cryptography = "^41.0.0"
sqlalchemy = "^2.0.0"
aiohttp = "^3.8.0"
python-dotenv = "^1.0.0"
qasync = "^0.24.0"
netifaces = "^0.11.0"
websockets = "^14.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 